adaptation to the new grace update
new shortcut where pressing '1' helps you deal with doomBringer, also made in a way so that you can use the modifier with pipebomb.

pressing '2' will only send 4, this is due to the plug modulers.

all these shortcuts are only for the modifiers if you don't use them, modify the script.